Serve models

+++ cnoms/server/server.py
@@ -4,6 +4,8 @@
     the actual flask app
 """
 import os, glob, json
+import os, glob
+from flask import Response
 from cnoms.models import Entry
 from cnoms.parser.parser import parse_html
 from cnoms import app
@@ -36,20 +38,34 @@ def get_field_history(user, site, fieldname=None):
 def show_template(user, site, template=None, edit=False):
     """render a template with the latest content for an entry"""
     entries = Entry.select().where(Entry.user==user, Entry.site==site).order_by(Entry.created.asc())
-    latest_entries = []
     seen_entries = []
+    simple_entries = []
+    collections = {}
+    items = {}
     for e in entries:
         if e.unique_id not in seen_entries:
-            latest_entries.append(e)
+            if e.type == "collection":
+                collections[e.fieldname] = []
+            elif e.type == "item":
+                collections.get(e.parent, {}).append(e.fieldname)
+                items[e.fieldname] = {}
+            elif e.parent:
+                items.get(e.parent, {})[e.fieldname] = e.value
+            else:
+                simple_entries.append(e)
             seen_entries.append(e.unique_id)
+
+    data = {entry.fieldname: entry.value for entry in simple_entries}
+    collections = {k: [items.get(item, {}) for item in kitems] for k, kitems in collections.items()}
+    data.update(collections)
+
     templates_path = app.jinja_loader.searchpath[0]
     if not template:
         template = 'index.html'
     current_template_path = os.path.join(templates_path, user, site, template)
     if not os.path.exists(current_template_path):
-        return 'template does not exist'
+        return Response(status_code=404)
     template_string = open(current_template_path).read()
-    data = {entry.fieldname: entry.value for entry in latest_entries}
     return render_template_string(template_string, __user=user, __site=site, cnoms_edit=edit, **data)
 
 @app.route('/<user>/<site>/<template>/edit')
@@ -73,11 +89,8 @@ def import_website(user=None, path_to_site=None):
     path = os.path.dirname(__file__)
     sitename = os.path.basename(os.path.normpath(path_to_site))
     new_static_path = os.path.join(path, '..', 'static', user, sitename)
-    # # for testign
-    # shutil.rmtree(new_static_path)
-    # shutil.rmtree(os.path.join(path, '..', 'static', user))
-    # shutil.rmtree(os.path.join(path, '..', 'templates', user))
-    shutil.rmtree(os.path.join(path, '..', 'static', user))
+    if os.path.exists(os.path.join(path, '..', 'static', user)):
+        shutil.rmtree(os.path.join(path, '..', 'static', user))
     if os.path.exists(os.path.join(path_to_site, 'static')):
         shutil.copytree(os.path.join(path_to_site, 'static'),
                         new_static_path)
@@ -98,11 +111,5 @@ def import_website(user=None, path_to_site=None):
                 f.write(str(template))
     print db_fields
     for db_entry in db_fields:
-        count = Entry.select().where(Entry.user==user,
-                                     Entry.site==sitename,
-                                     Entry.type==db_entry['type'],
-                                     Entry.value==db_entry['value'],
-                                     Entry.fieldname==db_entry['fieldname']).count()
-        if count == 0:
-            Entry.create(user=user, site=sitename, **db_entry)
+        Entry.get_or_create(user=user, site=sitename, **db_entry)
     return ''
